understanding how to check whether the tree is binary search tree

+++ binary_search/main.cc
@@ -9,29 +9,35 @@ int main(int argc, char *argv[])
     int idex;
     for (idex = 0; idex < 10; ++idex) {
         int data = rand()%100;
-        std::cout << data << std::endl;
         root = Insert(root, data);
     }
 
-    int min = GetMin(root);
-    int max = GetMax(root);
-    int height = GetHeight(root);
 
-    PrintBFS(root);
-    std::cout << std::endl;
-    PrintInOrder(root);
-    std::cout << std::endl;
+    if (IsBinarySearchTree(root)) {
 
-    std::cout << "Search 44 in the binary tree: "<< Search(root, 44) << std::endl;
+        int min = GetMin(root);
+        int max = GetMax(root);
+        int height = GetHeight(root);
 
-    std::cout << "Min: " << min << std::endl;
-    std::cout << "Max: " << max << std::endl;
-    std::cout << "Height: " << height << std::endl;
+        PrintBFS(root);
+        std::cout << std::endl;
+        PrintInOrder(root);
+        std::cout << std::endl;
 
-    DeleteValue(root, 44);
-    PrintInOrder(root);
-    std::cout << std::endl;
+        std::cout << "Search 44 in the binary tree: "<< Search(root, 44) << std::endl;
 
-    DeleteTree(root);
+        std::cout << "Min: " << min << std::endl;
+        std::cout << "Max: " << max << std::endl;
+        std::cout << "Height: " << height << std::endl;
+
+        DeleteValue(root, 49);
+        PrintInOrder(root);
+        std::cout << std::endl;
+
+        DeleteTree(root);
+    }
+    else
+        std::cout << "The tree is not binary search tree" << std::endl;
+
     return 0;
 }
